refactor: convert bootstrap API to async/await
