+++ src/core/components/bootstrap/list.js
@@ -1,0 +1,8 @@
+'use strict'
+
+module.exports = ({ repo }) => {
+  return async function list () {
+    const config = await repo.config.get()
+    return { Peers: config.Bootstrap || [] }
+  }
+}
