+++ src/core/components/stop.js
@@ -112,6 +112,11 @@ function createApi ({
 
   const api = {
     add,
+    bootstrap: {
+      add: Commands.bootstrap.add({ repo }),
+      list: Commands.bootstrap.list({ repo }),
+      rm: Commands.bootstrap.rm({ repo })
+    },
     config: Commands.config({ repo }),
     id: Commands.id({ peerInfo }),
     init: () => { throw new AlreadyInitializedError() },
